ДЗ08. Добавил генерацию идентификаторов для Book и BookComment; Доработал тесты;

+++ ru.otus.home-work-08/src/main/java/ru/otus/homework08/repository/BookRepository.java
@@ -3,7 +3,5 @@ package ru.otus.homework08.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ru.otus.homework08.domain.Book;
 
-import java.math.BigInteger;
-
-public interface BookRepository extends MongoRepository<Book, BigInteger> {
+public interface BookRepository extends MongoRepository<Book, Long> {
 }
